moved gotcourts API funcitonality to a separae class; code clean up and renaming
