+++ docker/Dockerfile
@@ -4,6 +4,6 @@ RUN mkdir /gotcourts
 WORKDIR /gotcourts
 ADD gotcourts gotcourts
 ADD setup.py setup.py
-RUN pip install . 
+RUN pip install .
 ENV PYTHONPATH=/gotcourts/gotcourts:$PYTHONPATH
-ENTRYPOINT [ "python", "/gotcourts/gotcourts/get_empty_slots.py" ]
+ENTRYPOINT [ "python", "/gotcourts/gotcourts/run.py" ]
