+++ setup.py
@@ -5,10 +5,10 @@ import setuptools
 
 setuptools.setup(
     name="gotcourts",
-    version="0.0.1",
+    version="0.1.0",
     # author="Example Author",
     # author_email="author@example.com",
     description="A small utility to list available slots on got courts website and send them to telegram channel",
     python_requires=">=3.6",
-    install_requires=["aiohttp", "pyyaml", "python-telegram-bot"]
+    install_requires=["aiohttp", "pyyaml", "python-telegram-bot"],
 )
